feat(as-api): wrap AssemblyScript PRNG exports within respective namespaces and only expose intended exports
Prevents AssemblyScript top-level state and functions from polluting global state for AS consumers.

BREAKING CHANGE: Only intended exports are now exposed through /assembly import rather than all source files, and consumers must now import respective namespaces
