+++ demo/pmc/pmc.js
@@ -9,8 +9,8 @@ import { WorkerBatchMode, WorkerSeedMode } from './enum.js';
 
 
 /* ******************************************************************************* *
-* Modify these parameters to change characteristics of the Monte Carlo simulation *
-* ******************************************************************************* */
+ * Modify these parameters to change characteristics of the Monte Carlo simulation *
+ * ******************************************************************************* */
 
 const TOTAL_POINT_COUNT = 1000000000;                   // 1 Billion points
 const PRNG = PRNGType.Xoshiro256Plus_SIMD;
