+++ src/index.ts
@@ -1,7 +1,10 @@
 /**
+ * PRNG JavaScript/TypeScript API - Can be used from most modern environments
+ * that support WebAssembly. See project README for compatability details.
+ * 
+ * This wrapper around the WebAssembly PRNGs simplifies WASM usage and makes 
+ * them thread safe from the JS runtime by allowing for separate WASM instances.
  * @packageDocumentation
- * JavaScript API - Can be used from any environment that supports WebAssembly.
- * See project README for compatability details.
  */
 
 export { PRNGType } from './types/prng';
